requirements.txt
diffusers
torch
subprocess32
discord.py
vertexai
fastapi
uvicorn
google-cloud-aiplatform